Better git checkout command sequence.

+++ lib/rails_pwnerer/app/vcs/git.rb
@@ -6,6 +6,7 @@ class RailsPwnerer::App::Git
   # remove any files not in Git in the application dir
   def cleanup_app_dir(app_name, instance_name, target_dir, app_name_is_dir = false)
     Dir.chdir(app_name_is_dir ? app_name : RailsPwnerer::Config[app_name, instance_name][:app_path]) do
+      next unless File.exist?(target_dir)
       Kernel.system "git clean -d -f -x -- #{target_dir}"
       Kernel.system "git checkout -- #{target_dir}"
     end
@@ -28,6 +29,7 @@ class RailsPwnerer::App::Git
   def revert_config_changes(app_name, instance_name)
     Dir.chdir RailsPwnerer::Config[app_name, instance_name][:app_path] do
       ['config', 'Gemfile', 'Gemfile.lock'].each do |dir|
+        next unless File.exist?(target_dir)
         Kernel.system "git clean -d -f -x -- #{dir}"
         Kernel.system "git checkout -- #{dir}"
       end
@@ -79,16 +81,8 @@ class RailsPwnerer::App::Git
     
     FileUtils.rm_rf app_path
     print "Doing Git clone, please enter your password if prompted...\n"
-    system "git clone #{git_repository} #{app_path}"
+    system "git clone -b #{git_branch} -- #{git_repository} #{app_path}"
     FileUtils.mkpath app_path unless File.exists? app_path
-    Dir.chdir app_path do
-      system "git checkout -q origin/#{git_branch}"
-      # NOTE: it is only safe to delete master now, we're on a remote branch
-      system "git branch -d master"
-      system "git checkout -q -b #{git_branch}"
-      system "git config branch.#{git_branch}.remote origin"
-      system "git config branch.#{git_branch}.merge #{git_branch}"
-    end
 
     # check that we really checked out a Rails app
     return check_rails_root(app_path) ? :ok : false
